spaces/skills-kalm/requirements.txt
streamlit>=1.40.0
sentence-transformers
torch
faiss-cpu
transformers
accelerate
huggingface_hub
